Change integer types

+++ src/lib.rs
@@ -29,17 +29,17 @@ use std::char;
 use std::cmp;
 use std::f32;
 
-static PIXEL_MAP: [[isize; 2]; 4] = [[0x01, 0x08],
-                                       [0x02, 0x10],
-                                       [0x04, 0x20],
-                                       [0x40, 0x80]];
+static PIXEL_MAP: [[u32; 2]; 4] = [[0x01, 0x08],
+                                   [0x02, 0x10],
+                                   [0x04, 0x20],
+                                   [0x40, 0x80]];
 
 /// A canvas object that can be used to draw to the terminal using Braille characters.
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct Canvas {
-    chars: HashMap<(usize, usize), isize>,
-    width:  usize,
-    height: usize,
+    chars: HashMap<(u32, u32), u32>,
+    width: u32,
+    height: u32,
 }
 
 impl Canvas {
@@ -47,7 +47,7 @@ impl Canvas {
     ///
     /// Note that the `Canvas` can still draw outside the given dimensions (expanding the canvas)
     /// if a pixel is set outside the dimensions.
-    pub fn new(width: usize, height: usize) -> Canvas {
+    pub fn new(width: u32, height: u32) -> Canvas {
         Canvas {
             chars: HashMap::new(),
             width: width / 2,
@@ -61,38 +61,38 @@ impl Canvas {
     }
 
     /// Sets a pixel at the specified coordinates.
-    pub fn set(&mut self, x: usize, y: usize) {
+    pub fn set(&mut self, x: u32, y: u32) {
         let (row, col) = (x / 2, y / 4);
         match self.chars.entry((row, col)) {
             Entry::Occupied(_) => {},
             Entry::Vacant(e) => { e.insert(0); },
         }
-        self.chars.get_mut(&(row, col)).map(|a| *a |= PIXEL_MAP[y % 4][x % 2]);
+        self.chars.get_mut(&(row, col)).map(|a| *a |= PIXEL_MAP[y as usize % 4][x as usize % 2]);
     }
 
     /// Deletes a pixel at the specified coordinates.
-    pub fn unset(&mut self, x: usize, y: usize) {
+    pub fn unset(&mut self, x: u32, y: u32) {
         let (row, col) = (x / 2, y / 4);
         match self.chars.entry((row, col)) {
             Entry::Occupied(_) => {},
             Entry::Vacant(e) => { e.insert(0); },
         }
-        self.chars.get_mut(&(row, col)).map(|a| *a &= !PIXEL_MAP[y % 4][x % 2]);
+        self.chars.get_mut(&(row, col)).map(|a| *a &= !PIXEL_MAP[y as usize % 4][x as usize % 2]);
     }
 
     /// Toggles a pixel at the specified coordinates.
-    pub fn toggle(&mut self, x: usize, y: usize) {
+    pub fn toggle(&mut self, x: u32, y: u32) {
         let (row, col) = (x / 2, y / 4);
         match self.chars.entry((row, col)) {
             Entry::Occupied(_) => {},
             Entry::Vacant(e) => { e.insert(0); },
         }
-        self.chars.get_mut(&(row, col)).map(|a| *a ^= PIXEL_MAP[y % 4][x % 2]);
+        self.chars.get_mut(&(row, col)).map(|a| *a ^= PIXEL_MAP[y as usize % 4][x as usize % 2]);
     }
 
     /// Detects whether the pixel at the given coordinates is set.
-    pub fn get(&self, x: usize, y: usize) -> bool {
-        let dot_index = PIXEL_MAP[y % 4][x % 2];
+    pub fn get(&self, x: u32, y: u32) -> bool {
+        let dot_index = PIXEL_MAP[y as usize % 4][x as usize % 2];
         let (col, row) = (x / 2, y / 4);
         let char = self.chars.get(&(row, col));
 
@@ -118,7 +118,7 @@ impl Canvas {
                 row.push(if char == 0 {
                     ' '
                 } else {
-                    char::from_u32((0x2800 + char) as u32).unwrap()
+                    char::from_u32(0x2800 + char).unwrap()
                 })
             }
             result.push(row);
@@ -131,7 +131,7 @@ impl Canvas {
         self.rows().into_iter().collect::<Vec<String>>().join("\n")
     }
 
-    fn line_vec(&self, x1: usize, y1: usize, x2: usize, y2: usize) -> Vec<(usize, usize)> {
+    fn line_vec(&self, x1: u32, y1: u32, x2: u32, y2: u32) -> Vec<(u32, u32)> {
         let xdiff = cmp::max(x1, x2) - cmp::min(x1, x2);
         let ydiff = cmp::max(y1, y2) - cmp::min(y1, y2);
         let xdir = if x1 <= x2 { 1 } else { -1 };
@@ -141,23 +141,23 @@ impl Canvas {
 
         let mut result = vec![];
         for i in 0..r + 1 {
-            let mut x = x1 as isize;
-            let mut y = y1 as isize;
+            let mut x = x1 as i32;
+            let mut y = y1 as i32;
 
             if ydiff != 0 {
-                y += ((i * ydiff) / r) as isize * ydir;
+                y += ((i * ydiff) / r) as i32 * ydir;
             }
             if xdiff != 0 {
-                x += ((i * xdiff) / r) as isize * xdir;
+                x += ((i * xdiff) / r) as i32 * xdir;
             }
 
-            result.push((x as usize, y as usize));
+            result.push((x as u32, y as u32));
         }
         result
     }
 
     /// Draws a line from `(x1, y1)` to `(x2, y2)` onto the `Canvas`.
-    pub fn line(&mut self, x1: usize, y1: usize, x2: usize, y2: usize) {
+    pub fn line(&mut self, x1: u32, y1: u32, x2: u32, y2: u32) {
         for &(x, y) in self.line_vec(x1, y1, x2, y2).iter() {
             self.set(x, y);
         }
@@ -201,13 +201,13 @@ impl Turtle {
     }
 
     /// Sets the width of a `Turtle`’s `Canvas`, and return it for use again.
-    pub fn width(mut self, width: usize) -> Turtle {
+    pub fn width(mut self, width: u32) -> Turtle {
         self.cvs.width = width;
         self
     }
 
     /// Sets the height of a `Turtle`’s `Canvas`, and return it for use again.
-    pub fn height(mut self, height: usize) -> Turtle {
+    pub fn height(mut self, height: u32) -> Turtle {
         self.cvs.height = height;
         self
     }
@@ -245,10 +245,10 @@ impl Turtle {
     /// brush is down.
     pub fn teleport(&mut self, x: f32, y: f32) {
         if self.brush {
-            self.cvs.line(cmp::max(0, self.x.round() as isize) as usize,
-                          cmp::max(0, self.y.round() as isize) as usize,
-                          cmp::max(0, x.round() as isize) as usize,
-                          cmp::max(0, y.round() as isize) as usize);
+            self.cvs.line(cmp::max(0, self.x.round() as i32) as u32,
+                          cmp::max(0, self.y.round() as i32) as u32,
+                          cmp::max(0, x.round() as i32) as u32,
+                          cmp::max(0, y.round() as i32) as u32);
         }
 
         self.x = x;
